GT2ModelTool: Implement JSON parsing for importing from OBJ. This means that the JSON file is now what is needed to convert from editable files, and denotes the OBJ to load. Remove the unknown data BIN file as it's now all mapped out in the JSON. Remove scale values from object names for 3DSMax compatibility and use the values in the JSON config file instead.

+++ GT2ModelTool/GT2ModelTool/ExportMetadata/MenuWheelMetadata.cs
@@ -6,9 +6,9 @@
         public ushort FrontWheelWidth { get; set; }
         public ushort RearWheelRadius { get; set; }
         public ushort RearWheelWidth { get; set; }
-        public int FrontLeftXOffset { get; set; }
-        public int FrontRightXOffset { get; set; }
-        public int RearLeftXOffset { get; set; }
-        public int RearRightXOffset { get; set; }
+        public short FrontLeftXOffset { get; set; }
+        public short FrontRightXOffset { get; set; }
+        public short RearLeftXOffset { get; set; }
+        public short RearRightXOffset { get; set; }
     }
 }
